+++ MatrixUtil/include/mutil/mat/mat.h
@@ -1,3 +1,11 @@
 #pragma once
 
 #include "mat_types.h"
+
+#include "matrix2.h"
+#include "matrix3.h"
+#include "matrix4.h"
+
+#include "intmatrix2.h"
+#include "intmatrix3.h"
+#include "intmatrix4.h"
